use bulk instead of bulk_index
In elasticsearch 1.7, there was of a function called bulk, which the aliased to bulk_index... for some reason.  In 1.9 they stopped aliasing it.  So in 1.7 you can import bulk or you can import bulk_index.  In 1.9 you can import bulk, but not bulk_index.  With this change, this will work with both 1.7 and 1.9.